Minibrowser animation no longer uses timers. We use .animate() callbacks instead.
Discussion with Lucas Garron over email:
The bug with the search panel is indeed odd. I believe its caused by a JS timer issue. We try to autofocus the search box for you, but if the search box is off-screen, then Firefox will force it onscreen and that messes up the animation. I recently improved the way the search panel works (it now uses CSS3 transitions and falls back to JS animation) so I'm going to remove the JS timer and use a callback that runs when the animation ends. That should fix the problem.